Further Completed Documents
Didn't fully complete everything, mainly the DFD, ERD, and a bit of the Use Case. The ERD can very much be improved but the DFD should be very much serviceable. The Presentation still is missing a bit and the use case is still missing the explanation of the use case.